+++ src/lib.ts
@@ -1,5 +1,5 @@
 import "./index.css";
 
-export { default as EasyEditor } from "./components/EasyEditor";
+export { default as EasyCodeEditor } from "./components/EasyCodeEditor";
 export { default as DefaultLight } from "./themes/DefaultLight";
 export { default as DefaultDark } from "./themes/DefaultDark";